requirements.txt
python-dotenv # a.k.a. dotenv, for reading environment vars from the ".env" file

#twint # for unoficially interfacing with twitter data
beautifulsoup4 # for scraping mobile Twitter site

google-cloud-bigquery # for interfacing with the BigQuery API
google-cloud-storage # for interfacing with Google Cloud Storage

sendgrid==6.0.5

pandas
networkx
psycopg2
sqlalchemy

# mpi4py # errors installing on heroku?
numpy
scipy

matplotlib




# formerly known as dev dependencies:
pytest
memory_profiler

# dev dependencies:
autopep8
